Reducing logic with passing tests.
